Fix const-cast in analyze_func
Summary:
We are getting closer to enabling memory compression for HHBBC. In this diff, we avoid the const-cast of the Func* in analyze_func; analysis is now const, as it logically should be. All code that updates a Func* (e.g. cfg-opts.h, dce.h, and optimize.h) now takes a php::MutFunc "deserialized function pointer" as well as the FuncAnalysis.

We still have to do a few const-casts in whole-program.cpp instead. These casts are shallow and easily to justify; with more work, I can eliminate them, but I'll do that in a follow-up.

My main concern with this diff is readability. It might be a good idea to bundle up {FuncAnalysis&, php::MutFunc} into a struct, since that's what code that optimizes funcs usually wants; we often have to update the func's bytecode and keep some analysis table in sync. The main annoyance with bundling is that some code is mutable in the func and const in the analysis, and some code is mutable in the func and takes ownership of (i.e. &&-ref) the analysis; I definitely don't want to proliferate these types. I'll try some of that in a follow-up as well.

Reviewed By: ricklavoie

Differential Revision: D23070609

fbshipit-source-id: 1d8e1e0757271779fede94382317990696cc89b0