Update Unify Error message to include backticks on types for markdown support
Summary:
This seems to be one of the most common error codes (4110 Unify Error), and since it includes code snippets in each message, it seems to be a good candidate for demonstrating markdown support in the new highlighted formatter.

Here's a few examples of the highlighted formatter:

{F278413517}

{F278413521}

{F278428934}

In this latter one, the subreason has a single-quoted, rather than backticked, code snippet, which is produced from `typing_reason.ml` (https://fburl.com/codesearch/ra4e96vw). Since changing that might produce a whole slew of additional changes, I'll put that in a separate diff.

Reviewed By: Wilfred

Differential Revision: D23043540

fbshipit-source-id: e26e0411a2843972c07cfcf799a5ee030c4f3ace