Merge ConstFunc and MutFunc
Summary:
These "fat pointers" are going to store a local copy of decompressed Func data. Copying this data would be expensive, so we delete the copy ctors on these structs.

By making these pointers no-copy, we get back some other ergonomics. As mofarrell and ricklavoie pointed out before, if the pointers are no-copy, we can use the const-ness of the pointer to decide the const-ness of the target, so we no longer need both ConstFunc and MutFunc. We merge them into a WideFunc type. We can also drop the blocks/blocks_mut distinction - it can now be inferred at the point of use.

WideFunc needs to have a static constructor for the const case - it's critical that if the user provides a const Func*, they get back a const WideFunc, and I don't think we can enforce that with a constructor. For parallelism, I created WideFunc::cns and WideFunc::mut, even though the mut case could be covered by a constructor.

Reviewed By: mofarrell

Differential Revision: D23077415

fbshipit-source-id: dc014c314a203a2847d3ddad2968c730060f8e19