Drop AnalysisContext from FuncAnalysisResult
Summary:
whole-program.cpp collects a FuncAnalysisResult for each func WorkItem it executes, or a list of these results for class WorkItem. If we had to store materialized Func bytecode on each of these results, we wouldn't save any memory.

Instead, we store a simple Context on FuncAnalysisResult. With a little bit of code motion, we don't need to add any new php::ConstFunc() constructor calls in this diff; every place that used the func's blocks from a FuncAnalysisResult also happens to have a materialized func around.

Reviewed By: mofarrell

Differential Revision: D23071295

fbshipit-source-id: 3a36cdd9a15d6fc86ba64017a685c76b35562a6d