Do not run full integrity check when looking up decls
Summary: This integrity check examines every pointer within a slab, ensuring that it falls within the slab and points to a valid value (i.e., a block preceded by a valid-looking header). But this traversal shouldn't strictly be necessary--if a slab contains the magic number which is written upon initialization, it should not be possible for its contents to be invalid. We should be able to assume that it's valid as long as the initialization constant is present, and avoid paying the overhead of this validation.

Reviewed By: arxanas

Differential Revision: D23039924

fbshipit-source-id: 999f8b08a7e219fec522762584e8725ae8e2a442