Add lightweight markdown-like syntax to error messages for highlighted formatter
Summary:
Adds support for a few markdown-like syntactic elements: `**bold**`, `*italics*` , and code with backticks (currently turns into underline). It currently doesn't allow nested delimiters, underscores, or tildes as delimiters, unlike normal markdown.

Since the main message in an error is already bolded, if a section in the main message is supposed to be bolded by being marked with ** **, I italicize bold it instead.

Note: I eventually want to get to storing and comparing literal output in review.sh/verify.py, since the .exp files don't show the formatting-related code, so it looks like they were never there to begin with in the .exp files.

Support for underline (__) or strikethrough (~~) is easy to add (see version 2 of this diff), but we had concerns about double underscores interfering with valid Hack identifiers like `__CLASS__`, plus we couldn't find a decent use case for strikethrough.

Reviewed By: Wilfred

Differential Revision: D22935740

fbshipit-source-id: e2d22b8efc8ccaaa4a615c3c30bfdbc470b32bb7