Analysis requires a WideFunc
Summary:
Right now, do_analyze takes a php::Func and constructs an internal php::WideFunc. That's safe and correct, but many callers already have a WideFunc. Since we're not planning to optimize the case of re-creating the wide pointer, let's allow the callee to pass one if they have it.

Cases we no longer need to deserialize:
- The initial analyze_func for optimization, in whole-program.cpp.
- The iterative analyze_func for optimization, in optimize.cpp.
- The extra analyze_func_inline for 86cinit, in analyze.cpp.

Cases where we now deserialize at the caller:
- All inline analyses from the index (makes sense - that's the prime case when analyzing one func we inspect another).
- The four calls to analyze_func used to build up a class analysis.
- Each const analysis in whole-program.cpp.

Reviewed By: ricklavoie

Differential Revision: D23084307

fbshipit-source-id: 7cc43366ad9d2a75fac87a2338781d6c6cdba8ea