Package up optimization context
Summary:
As mofarrell pointed out on a previous diff, now that optimization functions take both a FuncAnalysis and a WideFunc, we should to guarantee that the two match. We can do so and simplify our code at the same time by bundling optimization parameters into a common context type: I called it VisitContext, since we dispatch to these functions via visit_blocks.

VisitContext is the right type for global DCE, local DCE, inserting assertions, and optimizing iterators. We can ensure that the two Func pointers match at the point of construction.

Reviewed By: ricklavoie

Differential Revision: D23097441

fbshipit-source-id: 1a94e521a958c9eb8318f2222073ee77d0f7ea34