add self policy to function types
Summary:
Functions additionally need to have a self policy which guards to computational contents of the function. This allows us to implement `add_dependencies` for function types; it simply adds a flow into the function's self policy.

In addition, when a function is called we need to add a dependency from the function's self policy to the returned value. The pc of the called function is the upper bound of the current pc and the function's self policy.

Reviewed By: madgen

Differential Revision: D23030110

fbshipit-source-id: 46c478d1c2e5adb7b7aeacd0308382209a58e6af