do not create a hole for cipp calls
Summary: For cipp calls, we make assumptions about how all the inputs and outputs are policied and we therefore do not need to create a hole to globally solve the constraints introduced by cipp calls. In this diff, I directly create a prototype for the cipp function being called add a subtype constraint instead of a hole.

Reviewed By: madgen

Differential Revision: D23030876

fbshipit-source-id: 1c6ed7461e48e931966e638dbc8595a179e0d6f9