Burninate array AST nodes
Summary:
As title. These already raise parse errors, gated with a runtime option--let's
move to removing them entirely from the grammar.

Also fix the stragglers in the test suite.

Note that this degrades the error message a bit--it would be nice if I could
move to an intermediate state where only FFP has them and it is always an error
to run them through the lowerer, but we want to get here eventually anyways.

Reviewed By: Wilfred

Differential Revision: D22426429

fbshipit-source-id: e5c1022b7e334891a1832e986d42e7b0496f9080