Change strobelight/xenon backoff scheme
Summary:
Strobelight is configured currently to back off if it sees that Xenon has set a surprise flag in the past 10 seconds. This was based on a misunderstanding I had about how Xenon works.

While Xenon may not actually profile that often in a typical config, it's not because it isn't setting the surprise flag. The flag is configured to be set on a short period (e.g. every second), and it relys on the XenonRequestFreq to single out a small fraction of request that will be profiled. It is only on these specific requests that the surprise flag has any meaning for Xenon.

Because of this, I'm changing how Strobelight backs off.

The policy is now...

- If Strobelight is about to set a surprise flag, check m_isProfiledRequest, if true, disallow the write
- In event hooks, check m_isProfiledRequest, if true ==> Strobelight::active() == false
- if RuntimeOption::XenonForceAlwaysOn, we do the same as if m_isProfiledRequest is always true.

Reviewed By: shai-d

Differential Revision: D22718717

fbshipit-source-id: 4c46770d9c15b9106ab9b08a0983c21bf45b0c5c