Report data source and sinks for flow errors
Summary: What the title says. The idea is simple, we store position info. in ground policies, i.e., top, bottom, purpose which are the only things that survive simplification. This require inventing `Pos.none` in some places. This is uncomfortable, but in the absence of bugs, these shouldn't appear in violations.

Reviewed By: zilberstein

Differential Revision: D22999026

fbshipit-source-id: 831507a6a891c16c3ed02c581bc08bc7444b2a3b