Make policies carry multiple positions
Summary:
Allow policies to carry multiple positions. This is useful because we have join and meet defined on policies.

However, we can't currently write tests to expose this behaviour because for the moment all the joins and meets are on policy variables which this diff does not affect. This should save us from missing position bugs in the future.

Reviewed By: zilberstein

Differential Revision: D23104050

fbshipit-source-id: dc59457358523e5a76180c686b7546bbc732e6c0