Change how the top-level union is re-constituted when distributing intersections over unions
Summary: Distributing an intersection over unions gives back a list of types that need to be unioned together. Previously, that used a rather heavyweight operation, Typing_utils.fold_union, that tried to do full simplification. That should not be necessary (as long as we spot nothing, null, nested unions), since the components of the union are simplified, and don't need to be re-simplified. However, there is scope to also detect if some of the types being unioned are subtypes of others, if that turns out to be common.

Reviewed By: CatherineGasnier

Differential Revision: D23055744

fbshipit-source-id: a62e0d51ab56a698349b76faecb56401e80999c1