Expose abstract/final functions to userspace
Summary:
We need to implement `FactsAPI::isTypeAbstract()` natively:

https://www.internalfb.com/intern/diffusion/WWW/browsefile/master/flib/core/facts/facts.combined.php?lines=569

We've been saving this information to the DB all this time, but we've never stored it in-memory or returned it to users.

This diff adds two new functions:

```
namespace HH\Facts {

/**
 * True iff the given type cannot be constructed.
 */
function is_abstract(classname<mixed> $type): bool;

/**
 * True iff the given type cannot be inherited.
 */
function is_final(classname<mixed> $type): bool;

} // namespace HH\Facts
```

which expose this information to userspace.

Reviewed By: billf

Differential Revision: D23094593

fbshipit-source-id: 2abca8f34e8aac90005e45bc5602b536a122b7c2