Rename compression.h -> wide-func.h
Summary: This diff is pure code motion. We need a .cpp file to put the nontrivial implementation of WideFunc in. The right file to use is the one we've been calling compression.cpp, since whether or not we're using compression under the hood is irrelevant to the ser/de API.

Reviewed By: oulgen

Differential Revision: D23171069

fbshipit-source-id: 34c605bc47765a430b4e8ff3e59dc0a58a6ccd7c