option for file-backed cold arena
Summary: Cold and rarely changed data are put in the cold arena (e.g., bytecode). Here we create a temporary file and put the data on the file-backed mapping. The advantage is that we can easily move the data to disk when under memory pressure: the OS does it automatically, and the program can do it as well.  This would alleviate memory pressure and avoid OOMs.

Reviewed By: ottoni

Differential Revision: D23085550

fbshipit-source-id: 825b41a9d1c568ac7e461ac7d565085af80bd088