flow analysis for lambdas
Summary:
In this diff, I add support for analyzing lambdas.

The analysis is pretty straightforward. When we enter into a lambda, we freshen all of the local variables that get captured. This is because Hack uses function scoping, so flows into these variables inside the lambda cannot affect the versions of the variables from outside. We then proceed by creating a new pc, exception, and return types for the lambda which we replace inside the renv. Control flow inside the lambda has no effect outside of it, so we stash the renv before analyzing the inner statement block and then restore is afterwards.

Reviewed By: madgen

Differential Revision: D22950010

fbshipit-source-id: 1c1060f9d057d7a7e915edf8d7ad0d9bbf986e41