add ability to call any expression
Summary:
In order to call thngs that aren't top-level functions (eg, lambdas), we need to be able to call any expression.

Global functions and methods still need to be handled specially because we need to create holes when we invoke one. Local functions on the other hand do not need holes (even for global inference). Even if lambdas are passed between top-level functions, the constraints for them will be reflected in the decls of the functions that use them. So, all we have to do is ensure that the function being called is a subtype of the callsite type.

Reviewed By: madgen

Differential Revision: D22950002

fbshipit-source-id: 1980312cae97d9a0031a1406ec4e285b1dd28884