Remove SupportsAsyncEagerReturn flag from FCallArgs
Summary:
This flag is no longer used by anything, since return values of functions that
do not support async eager return became compatible with callers that request
async eager return.

The only useful information that hhbbc can deduce is whether a call never
supports async eager return, so that we can eliminate a bit check on the
return. But this is already representable by unsetting async eager target.

Note that the bit was shared with the LockWhileUnwinding flag, so its logic was
updated to occupy the original bit.

Reviewed By: oulgen

Differential Revision: D23182230

fbshipit-source-id: 11f5633d55d5f4f92c5ed86388209a2ee328eb55