Compat notice for throwing OOB on string index of varray
Summary: Emit compatibility notices when we throw an OOB exception accessing a string index of a varray (as opposed to the InvalidArgumentException thrown when this is done on a vec). This information will be used to correct code depending on this behavior and migrate varrays to also throw InvalidArgumentExceptions. The `array_exceptions` test has been updated to verify this behavior.

Differential Revision: D23189596

fbshipit-source-id: 267cdb6a432316a33c14fa6587118c201f3850ba