Normalize varray/vec, darray/dict exception messages
Summary:
Normalize some exception messages differing between dict/darray, and vec/varray. Unifying these messages as much as possible will make the migration to shared DArr/Dict and VArr/Vec IR ops simpler.

We can't totally unify the exceptions for dict/darray and vec/varray because there are certain places where the type of exception thrown varies. Specifically, when we get an element in a vec with a string key, we throw a InvalidArgumentException, while with a varray, we throw a OutOfBoundsException.

A new test (`fbcode/hphp/test/slow/array_exceptions.php`) was added to test this behavior explicitly.

Differential Revision: D23138576

fbshipit-source-id: 1243c58cc8a89d00582754e2b701295e7b96f2cd