Enable Func bytecode compression
Summary:
Now that all of our prep work is done, compressing a Func's bytecode is simple. We need to deserialize its bytecode when we create a WideFunc, and if it was created with mut(), we need to serialize the bytecode again on the way back.

In order to better handle traits, we store the Func's bytecode in a copy_ptr; by doing so, we can avoid duplicating the bytecode until we mutate it the first time. There are a number of other potential representations we could use for data "at rest", such as a combined copy_ptr-CompactVector type (with only one level of indirection), or a vector of copy_ptr-to-CompactVectors (in case we want partial sharing of bytecode across Funcs, even after they are edited), but I don't have convincing evidence yet that these alternatives are improvements. Right now, I have only a gross measurement of 20-30% memory savings.

As an immediate follow-up, we can delete the copy_ptr indirection used in BlockVec. I'll do that next.

Reviewed By: ricklavoie

Differential Revision: D23174785

fbshipit-source-id: 57f6ccc9c87b2edca48d55f8e4e14f69ecf87914