Ban __CLASS__ in class_meth on non final classes
Summary:
Using `__CLASS__` in `class_meth` technically does not do anything bad at runtime. It is the equivalent of `self::class`.

However, when we move to first class function pointer syntax, it will be difficult to up and replace these with their first class function pointer equivalents. Thus, we want to migrate off of this pattern, restrict the typechecker so that new ones are not added, to help migrate away from `class_meth`.

Reviewed By: periodic1236

Differential Revision: D22401546

fbshipit-source-id: f1b351935a72ee912e9da5c57fa6d6ab86e1c078