infrastructure for bounds on higher-kinded type parameters
Summary:
This is a step towards putting bounds on higher-kinded type parameters, which turned out to be convenient to have even for the simple version of higher-kinded types provided by this series of diffs.

Concretely, this changes `Type_parameter_env.tparam` such that type parameters have type parameters themselves. In a later diff `Type_parameter_env.tparam` will become the type of kinds in Hack.

This also affects the way the environment is queried for bounds on a type parameter. If `T` is a type parameter with n parameters itself, you must provide n type arguments when asking for `T`'s upper or lower bound. In the non-higher-kinded case, n is  always 0. Therefore, some interfaces in `Type_parameter_env` and `Typing_env` have changed accordingly.

Note that this does not implement the underlying logic for bounds on HK parameteres; it only provides the infrastructure for doing so later on.

Reviewed By: andrewjkennedy

Differential Revision: D23026932

fbshipit-source-id: ec233b20a42e4ffdbddf0230f2d6112fbcc9b6eb