add Typing_defs_core.Tunapplied_alias constructor
Summary:
This diff adds `Tunapplied_alias` as a constructor to `Typing_defs_core.ty_`.

This constructor is necessary to use type aliases in a higher-kinded fashion. Consider
```
type MyAlias1<T> = shape(‘key1’ => T);

type MyAlias2<T<_>> = T<int>;
```

Now, writing `MyAlias1<string>` is expanded transparently to the corresponding shape type. However, this is not possible if we want to use MyAlias1 as a higher-kinded type, for example by writing `MyAlias2<MyAlias1>`. Here, `MyAlias1` must not be expanded. However, we need a way of representing this deliberate usage of a non-fully-applied alias internally somehow.

This is where `Tunapplied_alias` comes into play: When an alias is used as a higher-kinded type, it does not get expanded, but instead turned into `Tunapplied_alias`.

Note that this diff only adds the constructor, the localization logic comes in a subsequent diff.

Reviewed By: andrewjkennedy

Differential Revision: D23027921

fbshipit-source-id: 893964e3d98dc0685be1c0f013f6af2cf1bd0db9