fix policy comparison in FlowSets
Summary:
When we added source locations to policies, we had to add logic to make comparisons ignore source locations. However, FlowSets were still using OCaml's built in structural equality. This led to bugs when examining flows for the same purpose policy in multiple places in the code.

It's hard to reproduce this without support for Public functions (coming in a diff soon) because we have a smart short-circuiting check in our `entailment_violations` function that just considers a flow to be ok if the policy names on the left and right are the same. I ran into the issue after trying to use a PUBLIC value inside a PUBLIC function. The analysis could not infer that `pc` flows into PUBLIC because the LHS of the entailment had the flow with a different source location. This diff fixed the issue

Reviewed By: madgen

Differential Revision: D23243157

fbshipit-source-id: b13d11c94ce3436080e355cf7faa73e719e08abd