Ban getting function pointers to abstract methods
Summary:
Getting a function pointer to an abstract static method should be banned in the typechecker.

Because we do not track the "abstract" property in the type checker, it is possible to call an abstract method with no type errors via `class_meth`.

```
<?hh

abstract class Foo {
  public abstract static function bar(): void;
}

function test(): void {
  $x = class_meth(Foo::class, 'bar');

  $x();
}
```
```
No errors
```

We want to make `class_meth` of an abstract method an error.

Reviewed By: Wilfred

Differential Revision: D22366240

fbshipit-source-id: 6138039bfbcc272c0f985f28baaefb2b1106cb01