Ensure typing environment is not dropped during subtyping
Summary:
Remove `invalid ()`, `valid ()`, and `default ()` in `typing_subtype.ml` to replace with versions that preserve the typing environment.

Historically, an invariant was maintained that subtyping did not introduce e.g. new type variables to the typing environment, and so dropping the environment between calls had no adverse effects.

Now, subtyping for the `Thas_member` constraint type calls out to `Typing_object_get.obj_get`, which can create fresh type variables through localisation in `Typing_phase`. Dropping the environment can therefore cause the typechecker to crash, when it tries to solve for a type variable which the environment does not know about.

Reviewed By: CatherineGasnier

Differential Revision: D23187927

fbshipit-source-id: 3e728ba737e8c628aad408ec841dccad7dfde6db