localization of higher-kinded types
Summary:
This is a major step towards higher-kinded types in Hack: Localization is now aware of the expected kinds of the type under consideration.

`Typing_phase.localize` is used on types that are supposed to be fully-applied (i.e., kind  `*`  ). Whenever we come across type arguments, we establish their expected kind either via the type parameter environment (in case of type parameters), or by translating the definitions of their type parameters into a kind (in case of clases, type aliases, etc). We then call `Typing_phase.localize_with_kind`, which is the counterpart to `localize` in case where we have an expected kind possibly other than *.

Crucially, Typing_phase does not report kind errors, this is done elsewhere (in a subsequent diff) for performance reasons. Therefore, the reasons of tracking kinds during localization are the following:
1) To determine whether an alias should be expanded or not (based on whether it is used as an HK type or not),
2) We return  `Terr` as the result of localization in case of kind mismatches where there isn't a sensible type to return (e.g., you supplied a fully-applied type in a place where a type constructor was expected). This is to reduce follow-up errors, so that the user can focus on fixing the original problem, rather then seeing a plethora of errors, obscuring the kind mismatch error.

Kinds are represented by two types:
1) `Typing_kinding_defs.kind` presents actual kinds, with all the necessary information
2) `Typing_kinding_defs.Simple.kind` is used as a simpler representation of kinds, in places where we do not care about checking type constraints (most importantly, during localization) but only about the "arities" of the involved types. In other words, `Typing_kinding_defs.Simple.kind` represents kinds that are built from `*` and `->`, but has no knowledge of bounds, which are present in the full representation of kinds, namely `Typing_kinding_defs.kind`.

Reviewed By: andrewjkennedy

Differential Revision: D23029936

fbshipit-source-id: aef02ee78edfbba2e15f99e4e8d58319f8f140af