subtyping of Tgeneric with type arguments
Summary:
This implements subtyping checks between on type parameters. Given that we do not allow constraints on them, they can only be a in a subtyping relationship if they use the same name `T`. Hence, we check `T<t_1, ..., t_n> <: T<t'_1, ... t'_m>` by checking that m = n and for all i in [1, n] we have t'_i <: t_i <: t'_i, as the type parameters of `T` can only be invariant at this time.
We do this by using the helper `simplify_subtype_variance`, which is also used for the same purpose for, say, subtype check between class types.

Reviewed By: andrewjkennedy

Differential Revision: D23132856

fbshipit-source-id: 35d1e5f1ac7b5a6f886c4baa20421ad3fffb3b04