remove --max-errors 1 from test/typecheck/instanceof tests
Summary: This removes  `--max-errors 1` from `hphp/hack/test/typecheck/instanceof/HH_FLAGS`. I'm doing this here because a latter diff in this series reorders the order in which some errors are reported for two files in `hphp/hack/test/typecheck/instanceof/`. With ` --max-errors 1` it looks as if the actual errors change, hiding the fact that there's actually just a reordering taking place.

Reviewed By: andrewjkennedy, Wilfred

Differential Revision: D23129581

fbshipit-source-id: 15bc33f01934e32677627f1bb5d4dc3106f20554