Mild refactor of function formatting
Summary:
The function for formatting fun-decl arguments had ballooned to handle the return type and where clauses, which made it harder to insert something in between. This separates them out and removes a useless value.

This has effect of decoupling the rule for splitting arguments from the rule for splitting returns. So, when a function return needs to split, the parameter list isn't also split as a consequence:

example diff
```
-  function f(
-  ): varray<shape(
+  function f(): varray<shape(
```

Reviewed By: arxanas

Differential Revision: D23250958

fbshipit-source-id: bcfaf3a2fc4019a986648547a8141a8cea7bace5