Option to ignore files with errors in comparator
Summary:
A big part of remaining divergences are due to differences in error recovery. For example, `param_mutability` is a single field decided based on presence of attributes. You are supposed to provide one of valid choices. But there is nothing stopping you from writing multiple ones:

```
 public function f(<<__Mutable, __OwnedMutable>> $x): void {
```
There will be an error, but declaration code will return a decl, choosing either `ParamOwnedMutable` or `ParamBorrowedMutable`, depending on ordering (and this ordering happens to be different between OCaml decl code and direct decl).

Maybe just to be sure we'll eventually want them too match too, but I think we should prioritize getting parity on valid code first, and this option allows to ignore such cases for now. This will allow us to land {D18710542} sooner and ensure that direct decl parser starts keeping up with changes to codebase.

Differential Revision: D23260041

fbshipit-source-id: 813acbb5ffc9bb8f8b0e762deb0ddda23456603a