Fix "Internal" serialization flavor to handle marked arrays correctly
Summary:
As title.

Previously, internal serialization was assuming "mark hack arrays" approach and produced marked vec/dict for x/X, which are not meant to exist at all with Eval.HackArrDVArrs=0.

This diff is a minimal change for "Internal" serialization to correctly preserve php/hack arrayness and marking with Eval.HackArrDVArrs on and off.

The are still more issues in VariableSerializer with Eval.HackArrDVArrs, but we'll fix them separately.

Also, my immediate motivation for making the change: I plan to reuse some of the "internal" serialization logic, for php_serialize "HAM-transparent" extension (unstable serialization format, which preserves both provenance tag and legacy bit).

Differential Revision: D23273483

fbshipit-source-id: cb0ad6c459c7fbddfed06f099ccc01640a198d8b