during naming, track only the names of type parameters in scope
Summary:
The module `Naming` currently has its own environment for type parameters in `genv.type_params` , of type
```
(Aast.reify_kind * (Ast_defs.constraint_kind * Aast.hint) list) SMap.t
```
However, carrying around the reification flag and constraints in this environment was only necessary to fill the `c_tparam_constraints` field of `Aast.class_tparams`.
The latter type is completely gone already, which means that there is no need anymore for `Naming` to carry around the reification flags and constraints. The only information that `Naming` needs are the *names* of the type parameters in scope, in order to decide which `Happly` nodes must be turned into `Habstr` (the hint equivalent of `Tgeneric`).

Therefore, this diff simplifies `Naming` such that it only carries a set of string to represent the type parameters in scope. This is just a refactoring and does not change observable behavior.

Reviewed By: Wilfred

Differential Revision: D23244187

fbshipit-source-id: e95c0b88df4af6cbdf92ccf6a3a592229d048854