remove useless Aast.class_tparams type
Summary:
This removes the type `Aast.class_tparams`, which was previously defined as follows:
```
and ('ex, 'fb, 'en, 'hi) class_tparams = {
  c_tparam_list: ('ex, 'fb, 'en, 'hi) tparam list;
  c_tparam_constraints:
    (reify_kind * (Ast_defs.constraint_kind * hint) list) SMap.t;
      [visitors.opaque]
      (** keeping around the ast version of the constraint only
       * for the purposes of Naming.class_meth_bodies
       * TODO: remove this and use tp_constraints *)
```
In practice, only the `c_tparam_list` field was still used, even though the `c_tparam_constraints` field was still filled, but not actually used for anything. I've therefore replaced all usages of this type with `tparam list`, which is the type used for type parameter lists of functions, methods, typedefs, etc.

This is a completely mechanical change.

Reviewed By: andrewjkennedy, Wilfred

Differential Revision: D23239279

fbshipit-source-id: e9539726faecff47fb97b8f30985f5dc9da419b6