Represent unescaped string literals with BString
Summary:
Hack strings are binary strings--there is no guarantee that they are valid UTF-8. However, we currently represent them in our Rust code with the String type, which is required to be valid UTF-8. Instead of using unsafe code to represent unescaped string literals with Strings (which is a violation of String's contract, and can cause panics and memory safety issues), this diff uses BString (a wrapper for Vec<u8>) to represent string literals.

The unsafe lines this diff adds in the emitter are not a safety regression--they are only a pushing-down of this safety issue to the next level. They'll be dealt with in a future diff.

Reviewed By: shiqicao

Differential Revision: D22020473

fbshipit-source-id: a72969d04e4526f17fbf48e97114015f569937ff