Simplify variance checks to be ready for intersection types
Summary: Well-formedness checks for generic parameters currently assume that method types are function types, and the code asserts if they're not. For an experimental treatment of multiple inheritance, method types may be intersection types. Simplify the variance checking to just call through to the checks for arbitrary types. This necessitates a small change to error messages for static methods; otherwise this is just a code cleanup.

Reviewed By: vsiles

Differential Revision: D23293900

fbshipit-source-id: 7ec952f83dccb984f0b335575fda46fd7043d507