Add new environment annotation statement
Summary:
In Hack IFC we will need to have precise type information at the beginning
of some blocks (e.g., loops and blocks guarded by is-checks). This diff
adds a new `AssertEnv` statement to store this information. Currently, it
is never generated in the annotated syntax tree.

I had to add a new `oxidized_hack` data type as using `'ex local_id_map`
directly broke the Rust tooling, I will check with the Rust crowd on how
to get rid of it.

Reviewed By: vsiles

Differential Revision: D23188195

fbshipit-source-id: 18a7faeb3dbdeda471e06b8495f4e01ce352f868