Cache NamedEntity on "symbol" strings
Summary:
We often need to do lookups from static strings to the classes, funcs, or TypeAliases that they name. In RepoAuthoritative mode, we should know the name of most symbols early on when we deserialize the autoload map. (Right now, that's done by the prod autoloader implementation, but the optimization could apply conceptually to the old PHP autoloader too.)

Designate all static strings that we allocate sufficiently early as "symbols", and pre-allocate space for a "SymbolPrefix" for these strings. This prefix contains a cached NamedEntity* and a cached Class*; in lowptr mode, it occupies 8 bytes.

On its own, this change is pretty clearly an instructions win, at least - around 0.5% on typical workloads. I tested two other versions of this change. In the first alternative, I preallocated this prefix for all static strings. I saw the instructions win, but I also saw a significant RSS regression - again around 0.5% - and the whole thing was a wash. In the second alternative, I squeezed a NamedEntity* into 3 bytes free in the static string via some hacks. I saw the same instructions win and no RSS regression; it also looked like a CPU win.

This version attempts to get the best of both worlds by identifying those strings which could benefit from the cache.

Reviewed By: oulgen

Differential Revision: D23283344

fbshipit-source-id: cc48d9c88592ecc02fe58da32a93ca18461e2703