reject certain unsupported classes and aliases as HK types
Summary:
The current implementation of higher-kinded types doesn't support using type constructors as higher-kinded types if they impose constraints on their input or output types. To this end, this diff detects two additional cases, where we must prevent certain types from being used as higher-kinded types. For illustration purposes, consider

```
class TakesHK<T<_>> {
  // T<int> used somewhere here
}
```

This diff ensures that the following two cases cannot be used as higher-kinded types:

1. Classes that impose bounds on their type arguments. Consider
`
class C<T as string> {...}
`
We must then reject `TakesHK<C>`, as `TakesHK` expects a more general takes constructor than `C` is. We achieve this by rejecting any usage of `C` as a higher-kinded types. Note that we need not reject classes with bounds on themselves. If we have
`
class D<T> as Foo<T> {}
`
then `TakesHK<D>` is okay.
In general, a constraint in a positive position (e.g., on the output of a type constructor) is okay, as the resulting kind is a subkind of the more general kind without a bound on the output.

2. Aliases do not allow specifying bounds on their type parameters. However, a alias may still *implicitly* impose bounds on its parameters. Consider
`class TakesNum<T1 as num> {}`
`type WithImplicitConstraint<T2> = TakesNum<T2>;`
Then we must still reject `TakesHK<WithImplicitConstraint>`, because implicitly, we still have `T2 <: num` as a bound on `T2` of `WithImplicitConstraint`.
To this end, this diff detects such aliases that implicitly impose constraints on their type parameters and rejects using these as higher-kinded types.

Reviewed By: andrewjkennedy

Differential Revision: D23168411

fbshipit-source-id: 31801064af31362fb8309f30b91655509cc80472